30th version of the HQS API
This is the 30th version of the HQS module. When tested under 60 conditions, it is a version that checks errors in all parts except median subtraction from the statistical calculation part, and corrections for mean, min, max, and percentage. In addition, only the comparison of the baseline of 80 people for the day of week is a version that has not yet been accurately implemented.